:sparkles: Now 32 bit system compatible

+++ src/lib/types/lang_type.rs
@@ -18,5 +18,5 @@ pub enum LangType {
     Eof,
 
     // for future Error Handling
-    Undefined(u64),
+    Undefined(u32),
 }
